+++ src/graphics/HISTORY.md
@@ -1,6 +1,11 @@
 Graphics Change History
 =======================
 
+@VERSION@
+------
+
+  * #701 Addressed rounding issue in svg implementation. This bug surfaces in edge cases of the PieChart in the charts module.
+
 3.10.1
 ------